requirements.txt
# Core Django
Django==4.2.16

# Database
psycopg2-binary==2.9.10

# Django Extensions
django-environ==0.11.2
django-extensions==3.2.3

# Rich Text Editor
django-ckeditor==6.7.0

# Internet Archive Integration
internetarchive==5.7.1

# Date Parsing
python-dateutil==2.9.0

# Image Processing (Python 3.13 compatible)
Pillow==11.0.0

# macOS Metadata (optional, only for development on macOS)
# Uncomment if running on macOS:
# osxmetadata==1.3.5

# Development Dependencies
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
